+++ src/utilities/index.js
@@ -1,7 +1,11 @@
 import calculateEVPercentage from './calculatePercentage';
+import getStateWinnerName from './getStateWinnerName';
+import getStateWinnerNames from './getStateWinnerNames';
 import getWinnerClassName from './getWinnerClassName';
 
 export {
   calculateEVPercentage,
+  getStateWinnerName,
+  getStateWinnerNames,
   getWinnerClassName
 };